// FixIt_FE/src/features/add-marketplace/type.ts
export type TAirsoftReplicas = {
  productName: string;
  gunCategory: string;
  customGunCategory: string;
  manufacturer: string;
  customManufacturer: string;
  gunType: string;
  condition: string;
  description: string;
  price: number;
  name: string;
  email: string;
  address: string;
  phone: string;
};